requirements.txt
streamlit
requests
numpy
scikit-learn
pyserial-asyncio
